BETA: Transformations rework: 	* Reworked mounts system as well: 		* Removed 4th custom mount type (That's transformation not a mount!) 		* Unhardcoded mount type. 		* Unhardcoded Ride packet. 	* Implemented retail like transformation system: 		* Replacing character's default stats. 		* Replacing character's default defenses. 		* Replacing character's default attacking type (Bow, Fist, Sword, etc..) 		* Unhardcoded water zone effect on transformations. 		* Unhardcoded ExBasicActionList packet (For transformations at least). 	* Thanks to: Nos 	* Reviewed by: Zoey76, Nos.

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/actor/stat/PlayableStat.java
@@ -239,9 +239,9 @@ public class PlayableStat extends CharStat
 	 * @return the base move speed of given movement type.
 	 */
 	@Override
-	protected double getBaseMoveSpeed(MoveType mt)
+	protected float getBaseMoveSpeed(MoveType mt)
 	{
-		double val = super.getBaseMoveSpeed(mt);
+		float val = super.getBaseMoveSpeed(mt);
 		
 		if (getActiveChar().isInsideZone(ZoneId.SWAMP))
 		{
